# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: Fantastic Coffee (decaffeinated)
  description: Fantastic Coffee (decaffeinated), the skeleton project for Web and Software architecture course
  version: 1.0.0
security:
  - bearerAuth: []
components:   
  securitySchemes:
    bearerAuth:     # arbitrary name for the security scheme
      description: "arbitrary name for the security scheme"       
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  schemas:
    ID: 
      title: General ID format 
      description: "Specifies the format of a unique identifier (UUID)."
      type: string 
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      minLength: 36
      maxLength: 36
      example: 3e7b82c1-8f66-4f54-9e57-7baa8a0ea8e3
               
               # 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
      readOnly: true
    Comment:
      title: Comment of a post
      description: |-
        comment provided by and user to an images of another user
      type: object 
      properties:
        # identifier of the post to which the comment belongs 
        post_id:
          $ref: "#/components/schemas/ID"

        # identifier of the comment/like in a post, given that we cann have many of the same user
        comment_id:
          $ref: "#/components/schemas/ID"

        # id of the user who does the comment/like, this will help us identify from whom the comment/like is
        user_id: 
          $ref: "#/components/schemas/ID"
          
        # the comment per se
        body:
          type: string
          description: "String representing a comment of a user"
          pattern: "^[a-zA-Z0-9]+$"
          minLength: 1
          maxLength: 2204
    Like:
      title: Like of a post
      description: |-
        like from a user to a photo/post of another user 
      type: object
      properties:
        # identifier of the post to which the like belongs  
        post_id:
          $ref: "#/components/schemas/ID"
        
        like_id:
          $ref: "#/components/schemas/ID"
        # id of the user who does the comment/like, this will help us identify from whom the comment/like is
        user_id: 
          $ref: "#/components/schemas/ID"
    Image:
      title: image of a post 
      description: |-
        An images regargly an user, such image will be the main component of an user post 
      type: object 
      properties:
        image:
          type: string
          description: "Binary representation of a image"
          format: base64
          pattern: '^[a-zA-Z0-9+/]*={0,2}$'
          minLength: 1  
          maxLength: 1048576
    Post:
      title: Post of an user on his profile 
      description: |- 
        a post make from an user that will be posted in his profile, this post contains an image which could have eventually likes and comment from others user 
      type: object
      properties:
        # identifier of the user whose did the post 
        user_id: 
          $ref: "#/components/schemas/ID"
        # identifier that will hepl us to identify the post of a user 
        post_id: 
          $ref: "#/components/schemas/ID" 
          
        # this specify the date in which the post was done and the time 
        uploaded:
          type: string
          description: "this specify the date in which the post was done and the time"
          format: data-time
          pattern: ^\d{4}-\d{2}-\d{2}\d{2}:\d{2}:\d{2}$
          maxLength: 19 
          minLength: 19
          example: 2022-01-01 12:34:56
        
        # the main object of this app are the image
        image:
          $ref: "#/components/schemas/Image"
        
        # collecttion of possible comments the post could have (i am not sure if thsi part have to be in the properties)
        comments:
          type: array
          description: "collecttion of possible comments the post could have (i am not sure if thsi part have to be in the properties)"
          items:  
            $ref: "#/components/schemas/Comment"
          minItems: 0
          maxItems: 2000
            
        # number of comments the post has 
        numberOfComments:
          type: integer
          description: "number of comments the post has"
        
        # collection of possible likes the post could have (i am not sure if thsi part have to be in the properties)
        likes:
          type: array
          description: "collection of possible likes the post could have (i am not sure if thsi part have to be in the properties)"
          items: 
            $ref: "#/components/schemas/Like"
          minItems: 0
          maxItems: 2000
            
        # number of likes the post could have
        numberOfLikes:
          type: integer
          description: "number of likes the post could have"
...
    User:
      title: User
      description: User schema.
      type: object
      properties:
        user_id:
          $ref: "#/components/schemas/ID"
        username:
          type: string
          description: "name of the user profile"
          pattern: '^.*?$' 
          example: "Daniel Guarnizo"
          minLength: 3 
          maxLength: 16         
    Profile:
      title: Profile of an user on the app 
      description: |-
        the profile for each user will contain information related to it, such as how many photos has been upload until now, how many follower and another user information
      type: object 
      properties: 
        user:
          $ref: "#/components/schemas/User"
        photos:
          type: array 
          description: an array of photos of a user
          items:
            $ref: "#/components/schemas/Post"
          minItems: 0 
          maxItems: 2200
            
        # indicates how many post/photos the user has in it's own profile
        numberOfPosts:
          type: integer
          description: "Number of posts/photos the user has in their profile."
          
        # number of follower indicates how many people follow this profile 
        userFollowers:
          type: integer
          description: "number of follower indicates how many people follow this profile "
          example: 2 # so this indicates that only two persons follow this user
        
        # number of other profiles the user follow 
        userFollowing:
          type: integer
          description: "number of other profiles the user follow "
          example: 1985 
    Stream:
      title: stream of a user
      description: |-
        the stream will show the post of all the following person of the user 
      type: object
      properties:
        # this wil be a collection of photos/post the user will see of all the others profiles his following 
        photos:
          type: array 
          description: "this wil be a collection of photos/post the user will see of all the others profiles his following "
          items:
            $ref: "#/components/schemas/Post"
          minItems: 0
          maxItems: 2000
...
paths:
...
  /users/{userid}:
    # using the patch method we are gonna to apply some modification to the resourrces without changing the complete resource 
    patch:
      security:
        - bearerAuth: []
      tags:
        - Profile
        - Users
      operationId: setMyUserName
      summary: User wants to change the username
      description: An user wants to change the username of its own profile.
      parameters:
        - name: userid
          in: path
          schema:
            $ref: "#/components/schemas/ID"
          required: true
      requestBody:
        description: New username for the user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newUsername:
                  type: string
                  description: "The new username to be set for the user."
      
      responses:
        "200":
          description: User update name was successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          description: User not found
        "500":
          description: Internal server error
    
  # UPLOAD PHOTO 
  # the user want to upload a post/photo to his profile 
...
  /users/{userid}/profile:
    get:
      security:
        - bearerAuth: []
      tags:
        - Users
        - Profiles
      operationId: getUserProfile
      summary: user search another user 
      description: |-
        an user search another user by name and try to get a possible resoult that match with his request 
      parameters:
        - name: userid
          in: path
          description: user id who is looking for another user 
          schema:
            $ref: "#/components/schemas/ID"
          required: true
          
        - name: username
          in: query
          description: username to search for 
          schema:
            type: string
            description: "username of the interesting profile"
            pattern: "^.*?$"
          required: true
          
      responses:
        # we use 200 instead of 201 because we are not creating a resource and we are just safistying a request
        "200":
          description: User profile found successfully 
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile"}

        '400':
            description: Bad request. Invalid user data.
            
        '401':
            $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: No user profiles found for the provided username.
        '500':
          description: Internal server error.
          
  # GET MY STREAM 
  /users/{userid}/stream:
  
    get:
      security:
        - bearerAuth: []
      tags:
        - Profile
        - Users
      operationId: getMyStream
      summary: Get the stream of an specific user
      description: |-
        an user wanna get the stream, so where all the post of the following users the user follow are posted.
      parameters:
        - name: userid
          in: path
          schema:
            $ref: "#/components/schemas/ID"
          required: true
      
      responses:
        "200":
          description: user stream with all the post of the followed users 
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stream"}
        '400':
            description: Bad request. Invalid user data.
            
        '401':
            $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: No user profiles found for the provided username.
        '500':
          description: Internal server error.